+++ instance/config.py
@@ -1,11 +1,13 @@
 class Config(object):
     DEBUG = False
+    DBNAME = "StackOverflow-lite"
 
 
 class DevelopmentConfig(Config):
     DEBUG = True
 
 class TestingConfig(Config):
+    DBNAME = "apptest_db"
     DEBUG = True
     TESTING = True
