+++ run.py
@@ -9,4 +9,3 @@ if __name__ == '__main__':
     database_connection = Dbcontroller()
     database_connection.create_tables()
     app.run()
-
